feat: 스켈레톤 애니 로직 훅화
