+++ src/components/ui/LoadingSkeleton.tsx
@@ -1,4 +1,5 @@
-import React, { useEffect, useRef } from 'react';
+import { usePulseAnimation } from '@/src/hooks/usePulseAnimation';
+import React from 'react';
 import { Animated, View } from 'react-native';
 
 interface LoadingSkeletonProps {
@@ -6,26 +7,7 @@ interface LoadingSkeletonProps {
 }
 
 export default function LoadingSkeleton({ count = 5 }: LoadingSkeletonProps) {
-  const fadeAnim = useRef(new Animated.Value(0.3)).current;
-
-  useEffect(() => {
-    const animation = Animated.loop(
-      Animated.sequence([
-        Animated.timing(fadeAnim, {
-          toValue: 1,
-          duration: 750,
-          useNativeDriver: true,
-        }),
-        Animated.timing(fadeAnim, {
-          toValue: 0.3,
-          duration: 750,
-          useNativeDriver: true,
-        }),
-      ]),
-    );
-    animation.start();
-    return () => animation.stop();
-  }, [fadeAnim]);
+  const fadeAnim = usePulseAnimation();
 
   return (
     <>
