+++ src/hooks/usePulseAnimation.ts
@@ -1,0 +1,27 @@
+import { useEffect, useRef } from 'react';
+import { Animated } from 'react-native';
+
+export const usePulseAnimation = () => {
+  const fadeAnim = useRef(new Animated.Value(0.3)).current;
+
+  useEffect(() => {
+    const animation = Animated.loop(
+      Animated.sequence([
+        Animated.timing(fadeAnim, {
+          toValue: 1,
+          duration: 750,
+          useNativeDriver: true,
+        }),
+        Animated.timing(fadeAnim, {
+          toValue: 0.3,
+          duration: 750,
+          useNativeDriver: true,
+        }),
+      ]),
+    );
+    animation.start();
+    return () => animation.stop();
+  }, [fadeAnim]);
+
+  return fadeAnim;
+};
